MentalHealth_RAG/MentalHealth/requirements.txt
streamlit
pypdf
langchain
sentence-transformers
langchain-community
opensearch-py
faiss-cpu

